requirements.txt
streamlit>=1.40.0
ollama>=0.1.0
requests>=2.28.0
python-dotenv>=1.0.0
